PRO-972: Extract CodeMirrorViewer to own file and export from sharedComponents
Moves the private CodeMirrorViewer component out of IssueDetailsModal.js
into its own file and exports it from src/sharedComponents/index.js so
the pro plugin can consume it via the @edac/shared-components external.
