+++ src/issueModal/components/IssueDetailsModal.js
@@ -11,6 +11,7 @@ import { useRef, useEffect, useState, useMemo } from '@wordpress/element';
 import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
 import IssueImage, { extractImageUrls } from './IssueImage';
+import CodeMirrorViewer from './CodeMirrorViewer';
 import FixPanel from './FixPanel';
 import DismissPanel from './DismissPanel';
 import Badge from '../../sidebar/components/Badge';
@@ -40,65 +41,6 @@ const getViewOnPageUrl = ( issue, viewLink ) => {
 	}
 
 	return url.toString();
-};
-
-/**
- * CodeMirror HTML viewer component
- *
- * @param {Object} props       - Component props.
- * @param {string} props.value - HTML code to display.
- */
-const CodeMirrorViewer = ( { value } ) => {
-	const textareaRef = useRef( null );
-	const editorRef = useRef( null );
-
-	useEffect( () => {
-		if ( ! textareaRef.current || ! window.wp?.codeEditor ) {
-			return;
-		}
-
-		// Initialize CodeMirror
-		const settings = window.wp.codeEditor.defaultSettings || {};
-		const editorSettings = {
-			...settings,
-			codemirror: {
-				...settings.codemirror,
-				mode: 'htmlmixed',
-				readOnly: true,
-				lineNumbers: true,
-				lineWrapping: true,
-				// Disable tab handling so it doesn't capture tab key
-				extraKeys: {
-					Tab: false,
-					'Shift-Tab': false,
-				},
-			},
-		};
-
-		editorRef.current = window.wp.codeEditor.initialize( textareaRef.current, editorSettings );
-
-		// Cleanup on unmount
-		return () => {
-			if ( editorRef.current?.codemirror ) {
-				editorRef.current.codemirror.toTextArea();
-			}
-		};
-	}, [] );
-
-	// Update content when value changes
-	useEffect( () => {
-		if ( editorRef.current?.codemirror ) {
-			editorRef.current.codemirror.setValue( value || '' );
-		}
-	}, [ value ] );
-
-	return (
-		<textarea
-			ref={ textareaRef }
-			defaultValue={ value }
-			className="edac-analysis__code-textarea"
-		/>
-	);
 };
 
 /**
